Ex22_Movie
Consider a Movie class that has members like, rating, name, year. Functions of the class:
public double getRating() 
public String getName()   
public int getYear() 

Suppose we wish to sort a list of Movies based on year of release. 
We can implement the Comparable interface with the Movie class, and we override the method compareTo() of Comparable interface.