Ex23_Movie2
Now, suppose we want to sort movies by their rating and names as well. When we make a collection element comparable(by having it implement Comparable), we get only one chance to implement the compareTo() method. The solution is using Comparator.

To compare movies by Rating, we need to do 3 things:
i. Create a class that implements Comparator (and thus the compare() method that does the work previously done by compareTo()).
class RatingCompare
class NameCompare
ii. Make an instance of the Comparator class.
iii. Call the overloaded sort() method, giving it both the list and the instance of the class that implements Comparator.